work on the submission for nft drop
